KL-149/style: separate filter method to util
